add edition manager singleton
The singleton is responsible for managing editions, that is the periods
of the goal-tracking by the app. The users starts an edition and their
performance in achieving their goals is being tracked by the app on a
daily basis, then the edition ends and a new one can begin. By default
editions last 365 days. They are all available through the
EditionsManager now implemented.