fix compound controls sharing state
When onSaveInstanceState is called, parcel with View's data is saved
with the View's ID, and onRestoreInstanceState is called for all views
with that ID with the data saved. Therefore if multiple instances of a
View share id (for example because multiple instances of a compound
control exist within the same layout and each has Views with same IDs as
every other instance of that compound control) then only the last
instance's state is persisted and then onRestoreInstanceState is
provided with that state for every instance: so all of them share the
state now. In order to fix this, the ReportForm was changed into a
Fragment.