let ReporterActivity properly interpret extras
When in extras Bundle, dayNumber is set to a day for which there already
are reports, the ReporterActivity will recognize that fact and load
those reports to populate it's controls with their data, ready to be
edited. If no reports exist for the day, on the other hand, it will
prepare clean form for brand new reports.