prevent data corruption when saving reports
An old method, which allowed saving a single Report to database, was
replaced by a new one, which only allows saving the exact amount of
Reports that is expected given the Edition. Additionally a method for
checking if a reports batch is valid, without crashing the app if it's
not, was added.

Several validation steps are included, ordered by execution cost,
ensuring that:

* the number of reports to be saved is the same as expected given Edition.

* all of the reports have timeStamps later than the Edition's start day.

* the entire batch of reports are for the same day.

* each report has a unique goal.

* all goals in reports are in fact goals of the Edition provided.

* no report already exists for the day.