add chart selection UI
A new Fragment was implemented, which presents a requested number of
buttons with one of them chosen as default (or none, if default value is
outside bounds of the provided array of button titles). It is now
plugged in the MainActivity and will be used to select chart type to be
shown.