fix leftover views after new data set
When setData() is called, SelectorButtonsFragment used to leave the old
buttons intact in the layout, creating the new ones on top of the old
set. This lead to uneven look of the semi-transparent parts of buttons,
and in case of not-matching number of views between setData calls, it
would leave the old buttons peaking between the new buttons.