fix activities having wrong parameters in manifest
MainActivity is now portrait only, as the UI doesn't make much sense in
landscape as it is now. A new layout can later be created to fit the
unique needs of the mode, but for now this is locked.

MainActivity can't be used in splitScreen feature too.

Both MainActivity and ReporterActivity are now launched in SingleTask mode