show daily notifications, handle clicks
A notification prompting the user to fill report form is now wired to
show on a daily basis, as well as after reboot, if needed. In case the
edition is ended already (reports are submitted on the next day, so this
is going to happen once every edition, at it's end) the new code is
smart enough to pick the latest rather than current edition, and add the
report to it.