improve report form layout
Fixed test overflowing past the end of the area intended for it. Fixed
scrollView being scrolled half the way down on startup on devices with
smaller screens. Removed old code which was trying to achieve the
former. Added full screen flag to remove the statusBar brief visibility
on activity's start.