smooth cumulative bar chart out
When a report is missing for a given day, it's corresponding bar on the
chart was not visible, since cumulative chart can get away with
repeating the same value when there is 0 difference, it will now display
the sum of values so far, as intended. Also margins between bars on the
chart were reduced for data sets with more than 48 items, as these
margins degrade visual quality of the chart with large sets.