add help feature
A help dialog, shown when "?" icons in MainActivity or in reports editor
are touched, was added. The dialog is showing help articles, which will
be expanded, past example text they are now, later on.

To make sure new dialog is of the same size as the reports editor
dialog, app's style was modified to define the desired dialog width,
which applies to both dialogs. Additionally EditorDialogFragment now
overrides onCreateView() (like the new HelpDialogFragment) rather than
onCreateDialog() as it used to. This is to keep them similar.

Help articles are parsed from string resources, with a simple parser
taking markup: title[p]paragraph text[/p]
and turning it into views with the text provided.