remove notification sound
Notifications are pushed during the night, at the first occation the
phone wakes up. Since they do not wake it up anyway and immediate action
of the user is not required, it is better for the notification to
remain silent. It is meant to be discovered early in the morning, not to
wake the phone and it's user up in the night. In order to achieve this,
IMPORTANCE_HIGH was changed to IMPORTANCE_LOW, which means the
notification will not peek and will not make sound.

Also removed a deprecated priority setting.