remove bugged notifications on click
When a notification is clicked, the app performs basic checks to make
sure it makes sense to respond to the click. For example if a report for
a given day already exists, the report form will not be opened again.
This poses a serious problem in cases where the notification is pushed
by mistake, because it is non-dismissable for the user. In case it's
refused to open the report form, the notification could stay there
indefinitely. Therefore a fix was issued in form of expanding upon the
old "just return" behavior in the receiver of the notification clicks.
Now it also removes the notification if it can't be used.

While it is not intended to erronously push a notification, it's better
to have this second layer of safety.