fix not saving goal name change
In ReporterActivity, when user chose to alter a goal's name, it used to
work only when filling a brand new report as opposed to editing an
existing one. The code responsible for saving the new goal name to
database was extracted and is now called from both the method which
saves a brand new report and the method which modifies an existing
report.