disable boot finished receiver when not needed
BoodFinishedReceiver will now only be enabled if an edition is going on.
To prevent potential bugs, it disables itself, when it runs and
encounters no ongoing edition. It is enabled by createEdition function
placed in EditionsManager object.

This way when the user no longer wants to actively use the app, it will
not run database queries at every reboot, nor run any other code.