make bar charts more readable
Bar charts now only show days within range of time passed, so after 10
days 10 bars are shown and they occupy all the space. Unlike before,
when they'd be tightly packed to leave room for the rest of the 365 days
total that will eventually be shown on the chart, as time passes.

The way scores for user reported "positives" are calculated in a new
way, using a max value and an asymptotic function with limit 1. The max
value is set to 40 and the function is 1 - e^-lambda*x. The new
calculations better serves the purpose of providing the user with
a rewarding yet challenging system.

Days with no reports now count towards weights applied to calculate
average daily efforts of the user. They are now correctly counted as
zeroes, which is much more intuitive than the old behavior.