fix issues with Spinner selection listener
Occasionally Spinner selection listener would get a null as a parameter
marked as not null. The parameter wasn't used anyway, so it was marked
nullable with no further changes to the code.

In MainActivity, when Spinner selection listener is invoked and says to
show data from a given edition, it is first chacked if the data is not
already being shown. If the data is already loaded, it is not loaded
again.