fix inactive listener seemingly being used
A listener reference was passed to a constructor, but then never used.
Instead an implementation of expected action was coincidentally present
withing the class that was listened to. This made it look like the
listening class was actually going to respond and it's implementation of
the interface was goint to be used, but in fact it was never used, it
just coincidentally was the same as the actuall hard coded response
within the other class.

The hard coded implementation was removed and the listener correctly
assigned and now put to use as intended.