code cleanup
Rewritten some of the code to leverage Kotlin's features and make it
more concise and readable.

Removed unused property in DaysData, along with it's value's eager
calculation. In future it might be re-added with lazy initialization
instead, for now it's never found no use for itself, so it was removed,
making the init block of the class much cleaner.

Removed unnecessary null checks in MainActivity, instead another
reference is used, which is quaranteed to not be null.