remove redundant backing property
An unnecessary backing property mData was removed as Kotlin can
auto-generate backing fields for properties. Therefore PieChart's data
property now has an auto-generated backing field and uses a default
getter to obtain value from it. Since value can be smart-cast to a
not-null type early in the setter, several instances of !! notation were
removed.

Additionally two other variables, to which values are assigned in the
data's setter, are no longer referenced other than to assign their new
values, which allowed for removal of further "!!" crash-if-null checks,
as these are not recommended for use, unless absolutely unavoidable.