color PieChart's slices according to scores.
Just like bars in the BarChart, now also PieChart's slices follow the
logic: first dimension of the score shown in size and second dimension
shown as color.

The method of generating the colors used was extraced from the BarChart,
so the same way of calculating the colors is used for both charts.

Unnecessary non selected slices color property was removed from PieChart, now
transparency is used for those slices instead.

DayData is now returning Scores objects instead of IntArrays, to provide
max values along with results.