use viewModel as data source for MainActivity
In order to achieve that EditionsManager was ported to the new
architecture, essentially being reflected in a repository form. DaysData
is no longer using the old constructor, so changes were made in old code
not to use it. This includes one case of feature disabling:
EditorDialogFragment is currently not operational - waiting for an
injection so it can obtain all the data it needs from a shared
ViewModel.