inject shared ViewModel in TopBarFragment
The fragment is now using MainActivityViewModel and data it holds,
instead of having it's own sources of data.
Also, a listener was removed from the fragment, because same
functionality is covered already by existing LiveData related code in
MainActivity. Hence there is no need to explicitly tell MainActivity to
update it's views.