fix validator taking report updates as duplicates
An update to a report was taken as a bug since only a single report per
goal per day is allowed. Therefore ReportsValidator was crashing the app
on update. It was now fixed by adding a parameter isEditing to the
validator's function isReportsBatchValid, when set to true, the check
against preexisting reports for a given day and goal are omitted.