code cleanup
Fixed a typo in filename of MainActivityViewModel.kt
Moved BroadcastReceiversModule to di package.
Moved BootFinishedReceiver to Alarms package, as this receiver is only
used to reset or do very similar stuff as AlarmsReceiver.
Moved several other helper classes to utils package to clean the main
directory up a bit further.